Move bed-specific code to BedBlockEntityTranslator.java
